requirements.txt
absl_py==1.3.0
numpy==1.23.5
Pillow==9.3.0
pytesseract==0.3.10
tensorflow==2.8.0 
tflite_model_maker==0.4.2
tflite_support==0.4.3
setuptools==58
pytqt5